requirements.txt
lightning_bolts==0.7.0
numpy==1.23
pandas==2.2.2
python-dotenv==1.0.0
pytorch_lightning==1.9.5
torchvision==0.11.2
tqdm>=4.66
wandb==0.17.2